pyproject.toml
[project]
name = "clinical_coder"
version = "0.1.0"
dependencies = [
  # Web Framework
  "fastapi==0.115.12",
  "uvicorn==0.34.2",
  # LLM & ML
  "langchain==0.2.17",
  "openai==1.78.0",
  "transformers==4.51.3",
  "torch==2.7.0",
  "sentence-transformers==4.1.0",
  "huggingface-hub==0.31.1",
  # Text Processing
  "nltk==3.9.1",
  "regex==2024.11.6",
  "rapidfuzz==3.13.0",
  "chardet==5.2.0",
  # Database
  "sqlite-utils==3.38",
  # Testing
  "pytest==8.3.5",
  "pluggy==1.5.0",
  "iniconfig==2.1.0",
  # Utilities
  "python-dotenv==1.1.0",
  "requests==2.32.3",
  "urllib3==2.4.0",
  "idna==3.10",
  "certifi",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = [
    "."
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
# uv automatically creates a lock section below; regenerate with `uv pip sync`
